+++ model.py
@@ -11,7 +11,7 @@ print __name__
 
 def build(P, input_size, embedding_size, controller_size, stack_size, output_size):
     softmax_output_size = output_size + 1
-    P.embeddings = np.random.randn(input_size + 2, embedding_size).astype(np.float32)
+    P.embeddings = np.random.randn(input_size + 2, embedding_size).astype(np.float64)
     controller_step = lstm.build_step(
         P, name="controller",
         input_size=embedding_size + stack_size,
@@ -22,15 +22,15 @@ def build(P, input_size, embedding_size, controller_size, stack_size, output_siz
     P.W_controller_output = 0.1 * np.random.randn(
         controller_size,
         softmax_output_size + stack_size + 1 + 1
-    ).astype(np.float32)
-    bias = np.zeros((softmax_output_size + stack_size + 1 + 1,), dtype=np.float32)
+    ).astype(np.float64)
+    bias = np.zeros((softmax_output_size + stack_size + 1 + 1,), dtype=np.float64)
     bias[-2] = 5
     bias[-1] = -5
     P.b_controller_output = bias
 
-    init_controller_cell = np.zeros((controller_size,), dtype=np.float32)
-    init_controller_hidden = np.zeros((controller_size,), dtype=np.float32)
-    init_stack_r = np.zeros((stack_size,), dtype=np.float32)
+    init_controller_cell = np.zeros((controller_size,), dtype=np.float64)
+    init_controller_hidden = np.zeros((controller_size,), dtype=np.float64)
+    init_stack_r = np.zeros((stack_size,), dtype=np.float64)
 
     def predict(ids, aux={}):
         X = P.embeddings[ids]
@@ -102,4 +102,4 @@ if __name__ == "__main__":
         inputs=[X],
         outputs=predict(X)
     )
-    print f(np.random.randn(10, 5).astype(np.float32))
+    print f(np.random.randn(10, 5).astype(np.float64))
